feat(phase8-9): scan-records dashboard + green CI + pinned lockfile
Phase 8 — dashboard/app.py rewritten to read scan_records: label counts
(locked/relative_value/reject), candidates table (observed gross gap vs fee-adjusted net
margin), rejections-with-reasons table, and the balance curve. Simulated-executable P&L is
shown as pending Phase 4/6b; nothing is presented as realized profit.

Phase 9 — make CI actually pass and pin deps:
- ruff now green on the CI command. Applied safe modernizations repo-wide (collections.abc
  imports, datetime.UTC, round() without redundant int(), zip(strict=False), dead import).
  Ignored UP042 (deliberate str+Enum) and RUF002/003 (em-dash/× in comments) in pyproject.
- requirements.txt: pinned lockfile via pip-compile (--extra dev). Generated on Python 3.13;
  regenerate on the 3.11 floor for production. CI keeps installing pyproject ranges on 3.11/3.12.

Lint auto-fixes are behavior-preserving (round() already returns int; strict=False is the
default). 72 tests pass; ruff check src tests scripts dashboard is clean.

+++ src/analysis/edge_calculator.py
@@ -63,7 +63,7 @@ class CrossExchangeEdge:
         notional = (self.leg_a_price + self.leg_b_price) * self.contracts
         if notional <= 0:
             return 0
-        return int(round(10_000 * self.net_usd / notional))
+        return round(10_000 * self.net_usd / notional)
 
 
 def cross_exchange_edge(
@@ -115,7 +115,7 @@ class BundleEdge:
         notional = sum(self.prices) * self.contracts
         if notional <= 0:
             return 0
-        return int(round(10_000 * self.net_usd / notional))
+        return round(10_000 * self.net_usd / notional)
 
 
 def bundle_long_edge(
